chore(): pull out header and input logic into new component
